move config file to utils
