+++ deploy/004-deploy-factories.ts
@@ -5,7 +5,7 @@ import { HardhatRuntimeEnvironment } from "hardhat/types";
 
 import { convertToUnit } from "../helpers/utils";
 import { ERC20__factory } from "../typechain";
-import { getConfig, getTokenConfig } from "./config";
+import { getConfig, getTokenConfig } from "../helpers/deploymentConfig";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts } = hre;
